+++ apps/web/src/_pages/space-search/constants.ts
