+++ apps/web/src/_pages/space-search/types.ts
