+++ apps/web/src/_pages/space-search/ui/space-card.tsx
