+++ apps/web/src/_pages/space-search/ui/space-search-header.tsx
